Fix Google Play Services NPE when scopes array has null entry
The logic on Android to filter out the 'email' scope will result in an
array of scopes that contains a `null` entry where the 'email' scope was
specified.

Google Play Services 11+ throws an null-pointer exception if it receives
an array of scopes containing a null.

This commit fixes the issue by using an ArrayList to collect non-'email'
scopes.